requirements.txt
flask>=3.0,<4.0
python-dotenv>=1.0,<2.0
requests>=2.31,<3.0
tenacity>=8.2,<9.0
mastercard-oauth1-signer>=1.2,<2.0
mastercard-client-encryption>=1.0,<2.0
PyJWT>=2.8,<3.0
